requirements.txt
netaddr
python-dotenv
requests
routeros-api
